requirements.txt
tensorflow
pandas
flask
hdfs
scikit-learn
pyshark
pyspark
kafka-python
chart-studio